services/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
python-multipart
httpx